add product ID in the safe name
